Split login runner into multiple files
