+++ cmd/login/kubecontext.go
@@ -1,0 +1,18 @@
+package login
+
+import (
+	"fmt"
+	"strings"
+)
+
+const (
+	contextPrefix = "gs-"
+)
+
+func generateKubeContextName(installationCodeName string) string {
+	return fmt.Sprintf("%s%s", contextPrefix, installationCodeName)
+}
+
+func isKubeContext(s string) bool {
+	return strings.HasPrefix(s, contextPrefix)
+}
